Allow non-default OldVersionHolder in a Jenkins pipeline
